% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_obj.R
\name{check_single_cell_object}
\alias{check_single_cell_object}
\title{check_single_cell_object}
\usage{
check_single_cell_object(obj, graph = NULL, reduction = NULL)
}
\arguments{
\item{obj}{An Seurat or SingleCellExperiment object to convert.}

\item{graph}{Name of the nearest-neighbour graph to use from seu@graphs
(e.g., "RNA_nn", "RNA_snn", or "SCT_nn").}

\item{reduction}{Name of the dimensionality reduction to use (default: "umap").}
}
\value{
A S3 scn_object with the following elements:
  \itemize{
    \item \code{type} - Character string: "Seurat" or "SingleCellExperiment"
    \item \code{embeddings} - Tibble of the selected reduction embeddings
    \item \code{key} - Character string of embedding dimension name key
    \item \code{metadata} - Data frame of cell metadata
    \item \code{graphs} - Seurat style nearest neighbour graph
    \item \code{n_cells} - Integer, number of cells in the dataset
    \item \code{cell_names} - Character vector of cell identifiers
  }
}
\description{
Takes a Seurat or SingleCellExperiment object and converts it into a scn_object.
}
\details{
Check Single-Cell Object Type and Extract Components
}
