% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{neighbour_distance.scaled}
\alias{neighbour_distance.scaled}
\title{neighbour_distance.scaled.}
\usage{
neighbour_distance.scaled(i, reduction, obj, graph = NULL)
}
\arguments{
\item{i}{The cell index.}

\item{reduction}{The name of the reduction map stored in the object to be used to calculate the coordinate variance.}

\item{obj}{A Seurat, SingleCellExperiment or SCNeighbours object containing single-cell data.
If in Seurat or SingleCellExperiment form will firist be converted to SCneighbours format}

\item{graph}{either a nearest neigbour graph in igraph, dgCMatrix or Seurat format, or the name of a graph stored in the Seurat object.
(e.g., "RNA_nn", "RNA_snn", or "SCT_nn").}
}
\value{
A number, the average variance in coordinates of neighbour cells of a certain cell i based on the provided reduction map.
}
\description{
The function to calculate the average variance in coordinates of neighbour cells of a certain cell i based on the provided reduction map.
}
