% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{calculate_neighbour_distance_for_all_cells}
\alias{calculate_neighbour_distance_for_all_cells}
\title{calculate_neighbour_distance_for_all_cells.}
\usage{
calculate_neighbour_distance_for_all_cells(seu, reduction, colname, graph)
}
\arguments{
\item{seu}{The Seurat object.}

\item{reduction}{The reduction map used to calculate the coordinate variance.}

\item{colname}{The column name to store the neighbourhood distance value in metadata.}

\item{graph}{Name of the nearest-neighbour graph to use from seu@graphs.}
}
\value{
A Seurat object with a new metadata column storing the variance in coordinates of neighbour cells for each cell
}
\description{
The function to calculate the average variance in coordinates of neighbour cells of all cells in the Seurat object based on the provided reduction map.
}
