% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{neighbour_distance.scaled}
\alias{neighbour_distance.scaled}
\title{neighbour_distance.scaled.}
\usage{
neighbour_distance.scaled(i, reduction, seu, graph = "RNA_nn")
}
\arguments{
\item{i}{The cell index.}

\item{reduction}{The reduction map used to calculate the coordinate variance.}

\item{seu}{The Seurat object.}

\item{graph}{Name of the nearest-neighbour graph to use from seu@graphs (default: "RNA_nn").}
}
\value{
A number, the average variance in coordinates of neighbour cells of a certain cell i based on the provided reduction map.
}
\description{
The function to calculate the average variance in coordinates of neighbour cells of a certain cell i based on the provided reduction map.
}
